check if client_id or client_secret are empty (#32)
* check if client_id or client_secret are empty

* removing the sunrise

* removing the sunrise

* adds a wait for the app update (it is only available when the installation is finished)

+++ install.sh
@@ -1,13 +1,17 @@
 #!/bin/bash
 
-docker stop nextcloud-container
+if [  -z $CLIENT_ID ] || [ -z $CLIENT_SECRET ]; then
+  echo "Please set CLIENT_ID and CLIENT_SECRET"
+  exit 1
+fi
+
+docker stop nextcloud-container || echo "No container to stop"
 sleep 1
 docker run -d --name nextcloud-container --rm --publish 80:80 nextcloud:28
-echo "Waiting for sunrise..."
 
 until docker exec --user www-data -it nextcloud-container php occ maintenance:install --admin-user=admin --admin-pass=admin >/dev/null
 do
-  echo "Try again waiting 2 seconds"
+  echo "Trying installation"
   sleep 2
 done
 
@@ -16,7 +20,13 @@ make appstore
 tar -xf ./build/artifacts/gdatavaas.tar.gz -C ./build/artifacts
 docker cp ./build/artifacts/gdatavaas nextcloud-container:/var/www/html/apps/
 docker exec -it nextcloud-container chown -R www-data:www-data /var/www/html/apps/gdatavaas
-docker exec --user www-data -it nextcloud-container php occ app:update --all
+
+until docker exec --user www-data -it nextcloud-container php occ app:update --all >/dev/null
+do
+  echo "Trying app update"
+  sleep 2
+done
+
 docker exec --user www-data -it nextcloud-container php occ app:enable gdatavaas
 
 docker exec --user www-data -it nextcloud-container php occ config:app:set gdatavaas clientId --value=$CLIENT_ID
